requirements.txt
gradio>=5.0,<6.0
# torch installed separately with cu130 index in Dockerfile
# diffusers installed from git main in Dockerfile (LTX2Pipeline)
transformers>=4.44.0
accelerate>=0.34.0
safetensors>=0.4.0
sentencepiece>=0.2.0
pydantic-settings>=2.0.0
httpx>=0.27.0
Pillow>=10.0.0
imageio[ffmpeg]>=2.34.0
av>=14.0.0
numpy<2.0
